server/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.6.0
asttokens==3.0.0
async-timeout==4.0.2
attrs==24.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
botocore==1.34.6
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constants==0.6.0
cryptography==40.0.2
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-shell-plus==1.1.7
django-simple-history==3.3.0
django-sse==0.4.1
djangorestframework==3.14.0
dynaconf==3.1.12
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.8
Flask==3.0.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.3.3
gevent==22.10.2
google-auth==2.28.2
google-auth-oauthlib==1.2.0
googlemaps==4.10.0
googlesearch-python==1.2.5
greenlet==2.0.2
gspread==6.0.2
h11==0.14.0
hvac==1.1.0
icecream==2.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.2
MouseInfo==0.1.3
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.0.0
pillow==10.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycares==4.3.0
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.18.0
pyhcl==0.4.4
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
redis==5.2.1
requests==2.32.3
requests-oauthlib==1.4.0
rsa==4.9
s3transfer==0.6.0
schedule==1.2.0
scipy==1.10.1
selenium==4.25.0
serpapi==0.1.5
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.1
sockets==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.25.0
SQLAlchemy==2.0.41
sqlparse==0.4.3
starlette==0.45.3
StrEnum==0.4.15
timezones==2.2.0
tkcalendar==1.6.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.34.0
validators==0.20.0
websocket-client==1.8.0
websockets==11.0.2
Werkzeug==3.0.1
wikipedia==1.4.0
win32-setctime==1.1.0
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
faiss-cpu
sentence-transformers